backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# AI and LangChain
langchain==0.0.350
langgraph==0.0.20
langchain-openai==0.0.2
langchain-anthropic==0.0.1
langchain-community==0.0.10
langsmith==0.0.69

# Vector database
pgvector==0.2.4

# Task queue and workers
celery==5.3.4
redis==5.0.1

# HTTP clients and API integrations
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Search and retrieval tools
tavily-python==0.3.1
serpapi==0.1.5
arxiv==2.1.0
crossref-commons==0.0.7

# Data processing and analysis
pandas==2.1.4
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0

# Document processing
beautifulsoup4==4.12.2
lxml==4.9.3
python-docx==1.1.0
python-pptx==0.6.23
reportlab==4.0.7
weasyprint==60.2

# Cloud storage
boto3==1.34.0
google-cloud-storage==2.10.0

# Observability and monitoring
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-exporter-otlp==1.21.0

# Security and validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7
rich==13.7.0
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
